每日填坑0601，儿童节快乐233，重要更新！Scheduler!!!!!!!!!!!

+++ src/main/java/cn/nukkit/scheduler/Task.java
@@ -21,7 +21,7 @@ abstract class Task {
         }
     }
 
-    public abstract void onRun(int currentTick);
+    public abstract void onRun(long currentTick);
 
     public void onCancel() {
